move disk history migrator into StorageMigrator
