+++ server/app/src/main/kotlin/com/vishnurajeevan/libroabs/graph/AppComponent.kt
@@ -3,8 +3,11 @@ package com.vishnurajeevan.libroabs.graph
 import com.vishnurajeevan.libroabs.App
 import com.vishnurajeevan.libroabs.StorageMigrator
 import com.vishnurajeevan.libroabs.libro.LibroApiHandler
+import com.vishnurajeevan.libroabs.models.Logger
 import com.vishnurajeevan.libroabs.models.graph.Named
+import com.vishnurajeevan.libroabs.models.libro.Book
 import com.vishnurajeevan.libroabs.models.server.ServerInfo
+import com.vishnurajeevan.libroabs.models.server.createPath
 import me.tatarka.inject.annotations.Provides
 import net.bramp.ffmpeg.FFmpeg
 import net.bramp.ffmpeg.FFmpegExecutor
@@ -12,6 +15,7 @@ import net.bramp.ffmpeg.FFprobe
 import software.amazon.lastmile.kotlin.inject.anvil.AppScope
 import software.amazon.lastmile.kotlin.inject.anvil.MergeComponent
 import software.amazon.lastmile.kotlin.inject.anvil.SingleIn
+import java.io.File
 
 @MergeComponent(AppScope::class)
 @SingleIn(AppScope::class)
@@ -61,5 +65,14 @@ abstract class AppComponent(
   @Provides
   @Named("parallel")
   fun parallelCount() = serverInfo.parallelCount
+
+  @Provides
+  @SingleIn(AppScope::class)
+  fun providesTargetDir(lfdLogger: Logger): (Book) -> File = { book ->
+    File("${serverInfo.mediaDir}/${book.createPath(serverInfo.pathPattern)}")
+      .also {
+        lfdLogger.log("Target Directory: $it")
+      }
+  }
 
 }
